+++ apps/psp/hasura/metadata/databases/default/tables/public_payment_providers.yaml
@@ -10,4 +10,4 @@ select_permissions:
     - logo
     - name
     filter: {}
-  role: ANONYMOUS
+  role: USER
