+++ apps/psp/src/pages/_app.tsx
@@ -1,32 +1,13 @@
-import { ApolloProvider } from '@apollo/client';
-import React, { useEffect, useState } from 'react';
-import createApolloClient from '../graphql/frontend/clients';
+import React from 'react';
 import '../index.css';
-import { UserProvider, useUser } from '../state/state';
-import { getUserFromToken } from '../utils/tokenUtils';
+import { UserProvider } from '../state/state';
+import ApolloWrapper from '../utils/ApolloWrapper';
 
 function App({ Component, pageProps }: any) {
   return (
     <UserProvider>
       <ApolloWrapper Component={Component} pageProps={pageProps}></ApolloWrapper>
     </UserProvider>
-  );
-}
-
-function ApolloWrapper({ Component, pageProps }) {
-  const [{ token }, dispatch] = useUser();
-  const [apolloClient, setApolloClient] = useState(createApolloClient());
-  useEffect(() => {
-    setApolloClient(createApolloClient(token));
-    const tokenFromLs = localStorage.getItem('token');
-    const user = getUserFromToken(tokenFromLs);
-    dispatch({ type: 'SET_TOKEN', payload: token });
-    dispatch({ type: 'SET_USER', payload: user });
-  }, [token]);
-  return (
-    <ApolloProvider client={apolloClient}>
-      <Component {...pageProps} />
-    </ApolloProvider>
   );
 }
